+++ inc/autoload.function.php
@@ -276,14 +276,6 @@ function glpi_autoload($classname) {
 
    $dir = GLPI_ROOT . "/inc/";
 
-   // Deprecation warn for Computer_Software* classes
-   if ($classname === 'Computer_SoftwareLicense') {
-      Toolbox::deprecated('Computer_SoftwareLicense has been replaced by Item_SoftwareLicense.');
-   }
-   if ($classname === 'Computer_SoftwareVersion') {
-      Toolbox::deprecated('Computer_SoftwareVersion has been replaced by Item_SoftwareVersion.');
-   }
-
    if ($plug = isPluginItemType($classname)) {
       $plugname = strtolower($plug['plugin']);