+++ inc/html.class.php
@@ -4472,36 +4472,6 @@ JS;
 
 
    /**
-    * Get javascript code for enable an item
-    *
-    * @param $id string id of the dom element
-    *
-    * @since 0.85.
-    * @deprecated 9.5.0
-    *
-    * @return String
-   **/
-   static function jsEnable($id) {
-      return self::jsGetElementbyID($id).".removeAttr('disabled');\n";
-   }
-
-
-   /**
-    * Get javascript code for disable an item
-    *
-    * @param $id string id of the dom element
-    *
-    * @since 0.85.
-    * @deprecated 9.5.0
-    *
-    * @return String
-   **/
-   static function jsDisable($id) {
-      return self::jsGetElementbyID($id).".attr('disabled', 'disabled');\n";
-   }
-
-
-   /**
     * Clean ID used for HTML elements
     *
     * @param $id string id of the dom element