+++ inc/itil_project.class.php
@@ -392,29 +392,4 @@ class Itil_Project extends CommonDBRelation {
       }
       echo '</div>';
    }
-
-   /**
-    * Duplicate all itil items from a project template to his clone.
-    *
-    * @deprecated 9.5
-    *
-    * @param integer $oldid  ID of the item to clone
-    * @param integer $newid  ID of the item cloned
-    *
-    * @return void
-    **/
-   static function cloneItilProject($oldid, $newid) {
-      global $DB;
-
-      Toolbox::deprecated('Use clone');
-      $itil_items = $DB->request(self::getTable(), ['WHERE'  => ['projects_id' => $oldid]]);
-      foreach ($itil_items as $data) {
-         unset($data['id']);
-         $data['projects_id'] = $newid;
-         $data                = Toolbox::addslashes_deep($data);
-
-         $itil_project = new Itil_Project();
-         $itil_project->add($data);
-      }
-   }
 }