+++ inc/projectteam.class.php
@@ -107,29 +107,6 @@ class ProjectTeam extends CommonDBRelation {
       }
    }
 
-   /**
-    * Duplicate all teams from a project template to his clone
-    *
-    * @deprecated 9.5
-    * @since 9.2
-    *
-    * @param integer $oldid        ID of the item to clone
-    * @param integer $newid        ID of the item cloned
-    **/
-   static function cloneProjectTeam ($oldid, $newid) {
-      Toolbox::deprecated('Use clone');
-      $team = self::getTeamFor($oldid);
-      foreach ($team as $type) {
-         foreach ($type as $data) {
-            $cd                  = new self();
-            unset($data['id']);
-            $data['projects_id'] = $newid;
-            $data                = Toolbox::addslashes_deep($data);
-            $cd->add($data);
-         }
-      }
-   }
-
 
    /**
     * Get team for a project